requirements.txt
websocket-client==1.4.2
Pillow==9.4.0
